+++ app/src/main/java/challenge/github/alc/com/popularmoveapp2/adapter/FavouritesMoviesAdapter.java
@@ -111,7 +111,7 @@ public class FavouritesMoviesAdapter extends RecyclerView.Adapter<FavouritesMovi
 
         //set the movie's rating
         favouriteMovie.setRating(cursor
-                .getLong(cursor.getColumnIndex(MovieContract.Favorites.FAVORITE_COLUMN_RATING)));
+                .getDouble(cursor.getColumnIndex(MovieContract.Favorites.FAVORITE_COLUMN_RATING)));
 
         //set the id
         favouriteMovie.setId(cursor
